Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    float x,i=1,mark;
+    int x,i=1,mark;
     scanf("%d",&x);
     mark=1<<(i-1);
     if(x & mark){
